feat: move ProxyServlet out from zimlet stack + cleanup
